+++ lib/ayadn/status.rb
@@ -454,7 +454,7 @@ module Ayadn
       end
     end
 
-    def has_to_migrate
+    def deprecated_ayadn
       say do
         @thor.say_status :deprecated, "Ayadn 1.x user data detected", :red
         @thor.say_status :warning,  "please delete your old ayadn folder then try again", :yellow